Code cleaning, good bye old leaderboard
